Extract wgpu-centric code from `app` into module `wgpu`
